requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
boto3==1.18.13
botocore==1.21.13
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.6
django-localflavor==3.1
django-storages==1.11.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jmespath==0.10.0
msgpack==1.0.2
mysqlclient==2.0.3
openpyxl==3.0.9
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-decouple==3.4
python-stdnum==1.16
pytz==2021.1
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
xlwt==1.3.0
zope.interface==5.4.0
